requirements.txt
ExamSpider
phthon-docx>=0.8.10
beautifulsoup4>=4.8.0
requests>=2.22.0

